+++ app/src/main/java/com/example/my_test_messenger/MainActivity.kt
@@ -8,6 +8,7 @@ import androidx.appcompat.widget.Toolbar
 import com.example.my_test_messenger.databinding.ActivityMainBinding
 import com.example.my_test_messenger.ui.ChatFragment
 import com.example.my_test_messenger.ui.SettingsFragment
+import com.example.my_test_messenger.ui.objects.AppDrawer
 import com.mikepenz.materialdrawer.AccountHeader
 import com.mikepenz.materialdrawer.AccountHeaderBuilder
 import com.mikepenz.materialdrawer.Drawer
@@ -22,9 +23,8 @@ import java.util.Objects
 class MainActivity : AppCompatActivity() {
 
     private lateinit var mBinding: ActivityMainBinding
-    private lateinit var mDrawer: Drawer
-    private lateinit var mHeader: AccountHeader
     private lateinit var mToolbar: Toolbar
+    private lateinit var mAppDrawer: AppDrawer
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -48,102 +48,15 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(mToolbar)
         supportFragmentManager.beginTransaction()
             .replace(R.id.dataContainer, ChatFragment()).commit()
-        createHeader()
-        createDrawer()
     }
 
-    private fun createDrawer() {
-        mDrawer = DrawerBuilder()
-            .withActivity(this)
-            .withToolbar(mToolbar)
-            .withActionBarDrawerToggle(true)
-            .withSelectedItem(-1)
-            .withAccountHeader(mHeader)
-            .addDrawerItems(
-                PrimaryDrawerItem().withIdentifier(100)
-                    .withIconTintingEnabled(true)
-                    .withName("Create Group")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_groups),
-                PrimaryDrawerItem().withIdentifier(101)
-                    .withIconTintingEnabled(true)
-                    .withName("Create Secret Chat")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_secret_chat),
-                PrimaryDrawerItem().withIdentifier(102)
-                    .withIconTintingEnabled(true)
-                    .withName("Create Channel")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_create_channel),
-                PrimaryDrawerItem().withIdentifier(103)
-                    .withIconTintingEnabled(true)
-                    .withName("Contacts")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_contacts),
-                PrimaryDrawerItem().withIdentifier(104)
-                    .withIconTintingEnabled(true)
-                    .withName("Calls")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_phone),
-                PrimaryDrawerItem().withIdentifier(105)
-                    .withIconTintingEnabled(true)
-                    .withName("Favorites")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_favorites),
-                PrimaryDrawerItem().withIdentifier(106)
-                    .withIconTintingEnabled(true)
-                    .withName("Settings")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_settings),
-                DividerDrawerItem(),
-                PrimaryDrawerItem().withIdentifier(107)
-                    .withIconTintingEnabled(true)
-                    .withName("Invite Friends")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_invate),
-                PrimaryDrawerItem().withIdentifier(108)
-                    .withIconTintingEnabled(true)
-                    .withName("Help")
-                    .withSelectable(false)
-                    .withIcon(R.drawable.ic_menu_help)
-            ).withOnDrawerItemClickListener(object : Drawer.OnDrawerItemClickListener {
-                override fun onItemClick(
-                    view: View?,
-                    position: Int,
-                    drawerItem: IDrawerItem<*>
-                ): Boolean {
-                    /**
-                     * Toast.makeText(applicationContext, position, Toast.LENGTH_SHORT).show()-  Уажет куда мы нажали
-                     */
-                    when (position) {
-                        7 -> supportFragmentManager.beginTransaction()
-                            /**
-                             * .addToBackStack(null) - дает возможность быстро вернуться к чатам
-                             */
-                            .addToBackStack(null)
-                            .replace(R.id.dataContainer, SettingsFragment()).commit()
-                    }
-                    return false
-                }
-            })
-            .build()
-    }
-
-    private fun createHeader() {
-        mHeader = AccountHeaderBuilder()
-            .withActivity(this)
-            .withHeaderBackground(R.drawable.header)
-            .addProfiles(
-                ProfileDrawerItem().withName("Kolo Toure")
-                    .withEmail("1223334444")
-            ).build()
-    }
 
     private fun initFields() {
         /**
          * инициализация
          */
         mToolbar = mBinding.mainToolbar
+        mAppDrawer = AppDrawer(this, mToolbar)
     }
 
 }
